chore: Extract `CMAccelerometerData+deviceOrientation`
